requirements.txt
Flask==2.3.3
pymysql==1.1.0
python-dotenv==1.0.0
cryptography==41.0.3
Werkzeug==2.3.7 